Refactor package names to lowercase
