+++ data/duke.txt
@@ -13,7 +13,6 @@ t false eat bread
 t false eat bread
 t false eat bread
 t false buy bread
-d false choose violence by Sep 9 23:59 2023
 t false eat bread
 t false buy bread
 t false eat bread
@@ -26,11 +25,14 @@ t false eat bread
 t false eat bread
 t false buy bread
 t false buy bread
-d false choose violence by Sep 9 23:59 2023
 t false choose violence
 t false eat bread
 t false eat bread
 t false buy bread
 t false buy bread
 t false deadline
-t false eat 
+t false eat
+t false choose violence
+d false eat bread by Sep 12 23:59 2023
+e false make bread from 2023-09-12T23:59 to 2023-09-13T23:59
+t false eat bread
