--- src/main/java/com/nyanbot/DukeGuiControllers/DukeLauncher.java
@@ -1,14 +1,0 @@
-package com.nyanbot.DukeGuiControllers;
-
-import javafx.application.Application;
-
-/**
- * Launches the GUI of the chatbot.
- *
- * @author Tan Kerway
- */
-public class DukeLauncher {
-        public static void main(String[] args) {
-            Application.launch(DukeMain.class, args);
-        }
-}
