--- src/main/java/com/nyanbot/DukeGuiElements/DukeDialogBox.java
@@ -1,69 +1,0 @@
-package com.nyanbot.DukeGuiElements;
-
-import java.io.IOException;
-import java.util.Collections;
-
-import com.nyanbot.DukeGuiControllers.DukeMainWindow;
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
-import javafx.fxml.FXML;
-import javafx.fxml.FXMLLoader;
-import javafx.geometry.Pos;
-import javafx.scene.Node;
-import javafx.scene.control.Label;
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import javafx.scene.layout.HBox;
-import javafx.scene.shape.Circle;
-
-/**
- * An example of a custom control using FXML.
- * This control represents a dialog box consisting of an ImageView to represent the speaker's face and a label
- * containing text from the speaker.
- *
- * @author Tan Kerway
- */
-public class DukeDialogBox extends HBox {
-    @FXML
-    private Label dialog;
-    @FXML
-    private ImageView displayPicture;
-
-    private DukeDialogBox(String text, Image img) {
-        try {
-            FXMLLoader fxmlLoader = new FXMLLoader(DukeMainWindow.class.getResource("/view/DialogBox.fxml"));
-            fxmlLoader.setController(this);
-            fxmlLoader.setRoot(this);
-            fxmlLoader.load();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
-        Circle clip = new Circle(Math.min(img.getHeight(), img.getWidth()));
-        displayPicture.setClip(clip);
-
-        dialog.setText(text);
-        dialog.setMaxHeight(Double.MAX_VALUE);
-        displayPicture.setImage(img);
-    }
-
-    /**
-     * Flips the dialog box such that the ImageView is on the left and text on the right.
-     */
-    private void flip() {
-        ObservableList<Node> tmp = FXCollections.observableArrayList(this.getChildren());
-        Collections.reverse(tmp);
-        getChildren().setAll(tmp);
-        setAlignment(Pos.TOP_LEFT);
-    }
-
-    public static DukeDialogBox getUserDialog(String text, Image img) {
-        return new DukeDialogBox(text, img);
-    }
-
-    public static DukeDialogBox getDukeDialog(String text, Image img) {
-        var db = new DukeDialogBox(text, img);
-        db.flip();
-        return db;
-    }
-}
