--- src/main/java/com/nyanbot/DukeTasks/InvalidTask.java
@@ -1,12 +1,0 @@
-package com.nyanbot.DukeTasks;
-
-/**
- * Encapsulates an invalid class.
- *
- * @author Tan Kerway
- */
-public class InvalidTask extends Task {
-    public InvalidTask(String errorMessage) {
-        super(errorMessage, true, false);
-    }
-}
