+++ src/main/java/com/nyanbot/dukeguicontrollers/DukeLauncher.java
@@ -1,0 +1,14 @@
+package com.nyanbot.dukeguicontrollers;
+
+import javafx.application.Application;
+
+/**
+ * Launches the GUI of the chatbot.
+ *
+ * @author Tan Kerway
+ */
+public class DukeLauncher {
+        public static void main(String[] args) {
+            Application.launch(DukeMain.class, args);
+        }
+}
