+++ src/main/java/com/nyanbot/duketasks/InvalidTask.java
@@ -1,0 +1,12 @@
+package com.nyanbot.duketasks;
+
+/**
+ * Encapsulates an invalid class.
+ *
+ * @author Tan Kerway
+ */
+public class InvalidTask extends Task {
+    public InvalidTask(String errorMessage) {
+        super(errorMessage, true, false);
+    }
+}
