--- text-ui-test/input.txt
@@ -1,41 +1,0 @@
-todo borrow book
-todo read book
-deadline return book /by this sunday
-list
-event project meeting /from Mon 2pm /to 4pm
-list
-deadline do homework /by no idea :-p
-mark 5
-list
-mark 1
-mark 2
-unmark 1
-list
-todo
-deadline
-event
-todo
-deadline do hw /by
-deadline do hw /by
-event do hw
-event do hw /from/to sunday
-event do hw /from sunday /to
-blah
-delete 3
-list
-delete -1
-delete 1000
-delete monkey
-mark -2.3
-mark 1000
-mark spamandeggs
-mark 1
-delete 2
-list
-delete 3
-delete 1
-delete 1
-delete 1
-mark 1
-unmark 1
-bye
